Splitting the  bookmark_via_url.py file in smaller tests
